feat: add Web3Provider
